requirements.txt
# Core
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4

# Computer Vision
opencv-python==4.9.0.80
pillow

# Geospatial
rasterio==1.3.10
geopandas==0.14.4
shapely==2.0.4
pyproj
fiona

# ML / YOLO
torch==2.2.2
torchvision==0.17.2
ultralytics==8.2.0

# Utilities
tqdm
PyYAML
requests
